reorganized directories. (EventsDataAccessObject.java shows errors due to local compatibility issue but no code was changed here)
