--- app/.gitignore
@@ -1,4 +1,0 @@
-# Ignore Gradle project-specific cache directory
-
-
-# Ignore Gradle build output directory
