+++ src/app/java/CalendarQuickstart.java
@@ -1,4 +1,4 @@
-package main.java;
+package app.java;
 
 import com.google.api.client.auth.oauth2.Credential;
 import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp;
