+++ src/app/resources/credentials.json
