+++ src/data_access/NotificationsDataAccessObject.java
@@ -5,7 +5,6 @@ import org.json.simple.JSONArray;
 import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 import org.json.simple.parser.ParseException;
-import use_case.NotificationsDataAccessInterface;
 
 import java.io.File;
 import java.io.FileReader;
