+++ src/entity/LocalCredentialsInterface.java
