+++ src/entity/LocalEventInterface.java
