+++ src/entity/NotificationsInterface.java
