+++ src/use_case/SignInOutInputBoundary.java
@@ -1,5 +1,8 @@
 package use_case;
 
+import InputData.LoginInputData;
+import OutputData.UserOutputData;
+
 public interface SignInOutInputBoundary {
     UserOutputData logIn(LoginInputData inputData);
     boolean logOut();
