+++ src/use_case/SignUpInputBoundary.java
@@ -1,5 +1,7 @@
 package use_case;
 
+import InputData.SignUpInputData;
+
 public interface SignUpInputBoundary {
     boolean createUser(SignUpInputData inputData);
 }
